requirements.txt
# Core AI Frameworks
langgraph==0.2.50
langchain==0.3.7
langchain-core==0.3.19
langchain-openai==0.2.8
langchain-community==0.3.7
langchain-google-genai==2.0.5
phidata==2.4.25

# LLM Integration
openai==1.54.0
anthropic==0.39.0
google-generativeai==0.8.3

# Vector Store
chromadb==0.5.3

# API & Web Framework
fastapi==0.115.4
uvicorn==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.1
python-multipart==0.0.9

# PDF Generation
reportlab==4.2.2
PyPDF2==3.0.1

# UI Framework
streamlit==1.38.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
aiofiles==24.1.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
